requirements.txt
# FastAPI & ASGI Server
fastapi~=0.115.8
uvicorn
# AI/NLP (For Task Parsing)
openai~=1.61.0
langchain
# Database (Optional)
sqlalchemy~=2.0.37
psycopg2  # If using PostgreSQL
psycopg2-binary  # If using PostgreSQL
# Background Tasks & Caching (Optional)
celery
redis
# Deployment & Observability (Optional)
gunicorn
python-dotenv~=1.0.1
pydantic~=2.10.6
pydantic-settings
asyncpg
alembic>=1.7.0
pgvector~=0.3.6
numpy~=1.26.4
requests~=2.32.3
setuptools~=69.0.3